Moved hero image into it's own component HeroImage, which is passed the character's image.
