+++ src/components/HeroImage/HeroImage.jsx
@@ -1,0 +1,30 @@
+import React from "react";
+import { s } from "./";
+
+export default function HeroImage(props){
+
+	const {
+		image: {
+			src,
+			mask: maskPath,
+			description
+		}
+	} = props;
+
+	// inline style to prevent the need for a huge png asset
+	const mask = { clipPath: `url(${maskPath})` };
+
+	return (
+		<div className={ s.hero } id="hero"> 	
+			<img 
+				className={ s.image }
+				src={ src } 
+				alt={ description }
+				style={ mask }
+			/>
+			<div className={ s.glow } id="glow"></div>
+		</div>
+	);
+}
+
+
