+++ src/components/HeroImage/index.js
@@ -1,0 +1,4 @@
+import s from "./styles.scss";
+
+export { s };
+export { default } from "./HeroImage.jsx";
