+++ src/components/HeroImage/styles.scss
@@ -1,0 +1,35 @@
+.hero {
+	position: fixed;
+	width: 100%;
+	height: 100%;
+	top: 0;
+	left: 0;
+	z-index: 0;
+}
+
+	.glow {
+		position: absolute;
+		width: 1em;
+		height: 1em;
+		top: 0;
+		left: 50%;
+		font-size: 100vw;
+
+	}
+		.glow::before {
+			content: "";
+			display: block;
+			width: 100%;
+			height: 100%;
+			background-color: yellow;
+			border-radius: 100%;
+			transform: translate(-50%, -50%);
+			filter: blur(0.2em);
+		}
+
+	.image {
+		display: block;
+		width: 100%;
+		height: 100%;
+		background-color: crimson;
+	}
